+++ hasklepias-main/examples/ExampleFeatures1.hs
@@ -19,19 +19,14 @@ module ExampleFeatures1
 import           ExampleEvents
 import           Hasklepias
 import           Test.Hspec
-
+import           Cohort.Attrition -- imported for test case
 {-
 Index is defined as the first occurrence of an Orca bite.
 -}
-indexDef
-  :: (Ord a)
-  => Definition
-       (Feature "events" (Events a) -> Feature "index" (Index Interval a))
-indexDef = defineA
-  (\events -> case firstConceptOccurrence ["wasBitByOrca"] events of
-    Nothing -> makeFeature $ featureDataL (Other "No occurrence of Orca bite")
-    Just x  -> pure (makeIndex $ getInterval x)
-  )
+defineIndexSet :: Ord a => Events a -> IndexSet Interval a
+defineIndexSet events = 
+  MkIndexSet $ fmap (setFromList . pure . makeIndex . getInterval) $
+    headMay $ makeConceptsFilter ["wasBitByOrca"] events
 
 {-  
 The baseline interval is the interval (b - 60, b), where b is the begin of 
@@ -197,8 +192,8 @@ type MyData
     , Feature "discontinuation" (Maybe (Int, Int))
     )
 
-getUnitFeatures :: Events Int -> MyData
-getUnitFeatures x =
+getUnitFeatures :: Index Interval Int -> Events Int -> MyData
+getUnitFeatures index x =
   ( idx
   , eval
     (buildContinuousEnrollment bline (containsConcepts ["enrollment"]) 60)
@@ -214,18 +209,21 @@ getUnitFeatures x =
   )
  where
   evs = pure x
-  idx = eval indexDef evs
+  idx = pure index
   bl  = fmap bline idx
   fl  = fmap flwup idx
 
+-- just a dummy set for now
+dummyIndex :: Index Interval Int
+dummyIndex = makeIndex $ beginerval 1 0 
 
 includeAll :: Events Int -> Criteria
 includeAll x = criteria $ pure
   (criterion (makeFeature (featureDataR Include) :: Feature "includeAll" Status)
   )
 
-testCohortSpec :: CohortSpec (Events Int) MyData
-testCohortSpec = specifyCohort includeAll getUnitFeatures
+testCohortSpec :: CohortSpec (Events Int) MyData Interval Int
+testCohortSpec = specifyCohort defineIndexSet includeAll getUnitFeatures
 
 example1results :: MyData
 example1results =
@@ -239,40 +237,25 @@ example1results =
   , pure $ Just (78, 18)
   )
 
-example2results :: MyData
-example2results =
-  ( makeFeature $ featureDataL (Other "No occurrence of Orca bite")
-  , makeFeature $ featureDataL (Other "No occurrence of Orca bite")
-  , makeFeature $ featureDataL (Other "No occurrence of Orca bite")
-  , makeFeature $ featureDataL (Other "No occurrence of Orca bite")
-  , makeFeature $ featureDataL (Other "No occurrence of Orca bite")
-  , makeFeature $ featureDataL (Other "No occurrence of Orca bite")
-  , makeFeature $ featureDataL (Other "No occurrence of Orca bite")
-  , makeFeature $ featureDataL (Other "No occurrence of Orca bite")
-  )
 
 exampleFeatures1Spec :: Spec
 exampleFeatures1Spec = do
 
   it "getUnitFeatures from exampleEvents1"
-    $          getUnitFeatures exampleEvents1
+    $          getUnitFeatures (makeIndex (beginerval 1 60)) exampleEvents1 
     `shouldBe` example1results
-
-  it "getUnitFeatures from exampleEvents2"
-    $          getUnitFeatures exampleEvents2
-    `shouldBe` example2results
 
   it "mapping a population to cohort"
     $          evalCohort testCohortSpec
                           (MkPopulation [exampleSubject1, exampleSubject2])
     `shouldBe` MkCohort
                  ( MkAttritionInfo 2 $ setFromList
-                   [ uncurry MkAttritionLevel (ExcludedBy (1, "includeAll"), 0)
-                   , uncurry MkAttritionLevel (Included                    , 2)
+                   [ uncurry MkAttritionLevel (SubjectHasNoIndex           , 1)
+                   , uncurry MkAttritionLevel (ExcludedBy (1, "includeAll"), 0)
+                   , uncurry MkAttritionLevel (Included                    , 1)
                    ]
                  , MkCohortData
-                   [ MkObsUnit "a" example1results
-                   , MkObsUnit "b" example2results
+                   [ MkObsUnit (makeObsID 1 "a") example1results
                    ]
                  )