+++ hasklepias-main/ChangeLog.md
@@ -1,4 +1,9 @@
-# Changelog for hasklepias
+# Changelog for hasklepias-main
+
+## 0.22.1
+
+* Updates examples to work with `hasklepias-core` `0.22.0`.
+* In preparation for a proper testing suite for the example app, moves the app part of the `exampleApp` code into to `Hasklepias.ExampleApp`.
 
 ## 0.22.0